processing/requirements.txt
requests
python-dotenv
beautifulsoup4
loguru
playwright
demoparser2==0.39.0
pandas
